% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphfoil_pkg.R
\docType{package}
\name{morphfoil}
\alias{morphfoil}
\title{morphfoil: Tools for morphing meshes in a .msh2 format}
\description{
The package contains a collection of functions which allow morphing
   and plotting of computational meshes in a .msh2 format for a flow over airfoils.
   Meshes in the .msh2 format are used by \code{xred} solver developed at MEiL PW.
}
\section{morphfoil functions}{

   \itemize{
     \item{\code{\link[morphfoil]{select_constr}}}
     \item{\code{\link[morphfoil]{select_morphp}}}
     \item{\code{\link[morphfoil]{morph_mesh}}}
     \item{\code{\link[morphfoil]{plot_mesh}}}
     \item{\code{\link[morphfoil]{plot_airfoil}}}
     \item{\code{\link[morphfoil]{read_mesh}}}
     \item{\code{\link[morphfoil]{write_mesh}}}
     \item{\code{\link[morphfoil]{read_airfoil}}}
   }
}

