% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mesh.R
\name{plot_airfoil}
\alias{plot_airfoil}
\title{Plot an airfoil}
\usage{
plot_airfoil(
  airfoil,
  morph_up,
  morph_lo,
  constr_up,
  constr_lo,
  air_points = TRUE
)
}
\arguments{
\item{airfoil}{A data frame containing coordinates of airfoil points.}

\item{morph_up}{A list of upper morphing points.}

\item{morph_lo}{A list of lower morphing points.}

\item{constr_up}{A list of upper constraints.}

\item{constr_lo}{A list of lower constraints.}

\item{air_points}{A logical value indicating if the airfoil
coordinate points should be plotted.}
}
\description{
The function plots an airfoil.
   The function uses \code{\link[graphics]{plot}},
   \code{\link[graphics]{lines}} and
   \code{\link[graphics]{arrows}} functions.
}
\examples{
# Select constraint points
constr_up <- select_constr(rae2822_airfoil, "up", rbind(c(0, 0), c(1, 1)))
constr_lo <- select_constr(rae2822_airfoil, "lo", rbind(c(0, 0), c(1, 1)))

# Select morphing points
x_coord <- seq(0.15, 0.9, len = 5)
morph_up <- select_morphp(rae2822_airfoil, "up", x_coord)
morph_lo <- select_morphp(rae2822_airfoil, "lo", x_coord)

plot_airfoil(rae2822_airfoil, morph_up, morph_lo, constr_up, constr_lo)

}
