% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mesh.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a mesh}
\usage{
plot_mesh(mesh, airfoil, xl = c(0, 1), yl = c(-0.2, 0.2))
}
\arguments{
\item{mesh}{A list of elements \code{$head}, \code{$coord} and \code{$tail}.}

\item{airfoil}{A data frame containing coordinates of airfoil points.}

\item{xl}{A vector of limits for the x-axis.}

\item{yl}{A vector of limits for the y-axis.}
}
\description{
The function plots a mesh in the .msh2 format.
   The function uses \code{\link[graphics]{plot}} and
   \code{\link[graphics]{lines}} functions.
}
\examples{
plot_mesh(rae2822_mesh, rae2822_airfoil)

}
