% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rae2822_airfoil}
\alias{rae2822_airfoil}
\title{Coordinates of the RAE 2822 airfoil}
\format{
A data frame with 65 rows and three variables.
\describe{
  \item{x}{x coordinates of points}
  \item{yup}{y coordinates of the upper points}
  \item{ylo}{y coordinates of the lower points}
}
}
\source{
\url{https://www.grc.nasa.gov/WWW/wind/valid/raetaf/geom.txt}
}
\usage{
rae2822_airfoil
}
\description{
The RAE 2822 airfoil coordinates are stored in a three-column format.
A length of the airfoil chord is normalized to one.
}
\keyword{datasets}
