% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rae2822_mesh}
\alias{rae2822_mesh}
\title{Mesh in the .msh2 format for the RAE 2822 airfoil}
\format{
A list with three elements:
\describe{
  \item{head}{A character string containing basic information about a file}
  \item{coord}{A data frame containing points coordinates}
  \item{tail}{A character string describing volumes and boundaries}
}
}
\usage{
rae2822_mesh
}
\description{
The mesh is unstructured an is suitable for a \code{xred} solver
developed at MEiL PW.
}
\keyword{datasets}
