% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_mesh.R
\name{read_airfoil}
\alias{read_airfoil}
\title{Read a file with airfoil coordinates}
\usage{
read_airfoil(name, skip_rows = 5, ...)
}
\arguments{
\item{name}{A string with a name of the file to read.
The file name is relative to a working directory.}

\item{skip_rows}{A number of rows to skip before reading a table.
The default value is 5.}

\item{...}{additional argumets passed to \code{read.table}.}
}
\value{
The output data frame contains: x coordinates of points (\code{$x}),
   y coordinates of the lower points (\code{$ylo}) and
   y coordinates of the upper points (\code{$yup}).
}
\description{
The function reads a file in a three-row format
   and store it in a data frame.
   The function uses \code{\link[utils]{read.table}} function.
}
