% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_mesh.R
\name{read_mesh}
\alias{read_mesh}
\title{Read a file in a msh2 format}
\usage{
read_mesh(name, head_rows = 4)
}
\arguments{
\item{name}{A string with a name of the file to read.
The file name is relative to a working directory.}

\item{head_rows}{A number of rows containing introductory information.
The default value is 4.}
}
\value{
The output list contains: a head of the file (\code{$head}),
mesh points (\code{$coord}) and a file tail (\code{$tail}).
}
\description{
The function reads a .msh2 file and store it in a list.
   The function uses \code{\link[base]{scan}} and
   \code{\link[utils]{read.table}} functions.
}
