% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_mesh.R
\name{select_constr}
\alias{select_constr}
\title{Select points constraining morphing}
\usage{
select_constr(airfoil, surf, limits)
}
\arguments{
\item{airfoil}{A data frame containing coordinates of airfoil points.}

\item{surf}{A character string indicating upper or lower curve
of the airfoil.}

\item{limits}{A matrix which in each row contains limits within wich
the constraint points should lie.}
}
\value{
A list of two elements: a logical vector (\code{$sel}) indicating
   which points are selected and a data frame (\code{$coord}) containing
   coordinates of the selected points.
}
\description{
The function selects points from airfoil points which
should not be moved during morphing of a mesh.
}
\examples{
constr_up <- select_constr(rae2822_airfoil, "up", rbind(c(0, 0), c(1, 1)))
constr_lo <- select_constr(rae2822_airfoil, "lo", rbind(c(0, 0), c(1, 1)))

}
