% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_mesh.R
\name{write_mesh}
\alias{write_mesh}
\title{Write a file in a msh2 format}
\usage{
write_mesh(name, mesh)
}
\arguments{
\item{name}{A string naming a file to write to.
The file name is relative to a working directory.}

\item{mesh}{A list of three elements: a head of the file (\code{$head}),
mesh points (\code{$coord}) and a file tail (\code{$tail}).}
}
\description{
The function writes a .msh2 file.
   The function uses \code{\link[base]{write}} and
   \code{\link[utils]{write.table}} functions.
}
