% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphfoil_pkg.R
\docType{package}
\name{morphfoil}
\alias{morphfoil}
\title{morphfoil: A package for morphing meshes in a .msh2 format}
\description{
The package contains a collection of functions for morphing
   computational meshes in a .msh2 format for a flow over airfoils.
   Meshes in the .msh2 format are used by xred solver developed at MEiL PW.
}
